+++ PA9.4/Player.h
@@ -5,7 +5,7 @@ public:
 	Player() { life = 10; speed = 150; }
 
 	void update(sf::FloatRect& bounds, 
-		vector<Bullet*>& pBull, sf::RenderWindow* window, double& dt);
+		vector<Bullet*>& pBull, sf::RenderWindow* window, float& dt);
 
 
 private:
